Halve physics memory (float32) to fix MemoryError at finer grids
lfilter/np.gradient silently upcast to float64 -> ~1.76 GB/array at 1.0deg -> crash.
Now: charge cast to float32; gradient computed in float32 + in-place (lower peak);
epsilon cast to float32 so E stays float32; run_pipeline(keep_anom=False) frees the
anomaly cube during sweeps. Results identical (7421 events, recall unchanged).

+++ src/climate_capacitor/physics/charge.py
@@ -37,7 +37,9 @@ def accumulate_charge(
     # Causal FIR: y[t] = sum_k kernel[k] * x[t-k]. NaNs are treated as 0 so
     # short gaps don't blow up the accumulation (logged upstream in Phase 2).
     data = np.nan_to_num(anomaly.values, nan=0.0)
-    charged = lfilter(kernel, [1.0], data, axis=axis)
+    # lfilter upcasts to float64; cast back to float32 to halve memory (matters
+    # at finer grids where a single cube is ~1 GB).
+    charged = lfilter(kernel, [1.0], data, axis=axis).astype(np.float32, copy=False)
     out = anomaly.copy(data=charged)
     out.name = "charge"
     out.attrs["long_name"] = "accumulated thermal charge"
